Print test summary including failed test's result
h2spec prints out the progress of tests beautifully, but most of the
times we are interested in failed tests.  Currently, we have to look
carefully the terminal while tests are running or scroll back the
terminal to see whether failed tests exist or not.

To solve the above problem, this commit adds summary output after all
tests are performed.  It prints out the failed test cases including
expected and failed pair in a same format shown during testing.  It
also prints out the number of tests, passed, skipped and failed test
cases so that we can see what happened easily.